Consolidate duplicated constants into @boardsesh/board-constants
HOLD_STATE_MAP (4 copies), grade colors (2 copies), STATE_TO_CODE / role
codes (2 copies), and convertLitUpHoldsStringToMap (2 copies) are now
each defined once in board-constants. Web and backend re-export for
backwards compatibility. Also adds AUX to the shared HoldState type.
