+++ packages/backend/src/db/queries/util/hold-state.ts
@@ -1,124 +1,2 @@
-import type { BoardName } from '@boardsesh/shared-schema';
-import type { LitUpHoldsMap, HoldState } from '@boardsesh/shared-schema';
-
-type HoldColor = string;
-type HoldCode = number;
-
-export const HOLD_STATE_MAP: Record<
-  BoardName,
-  Record<HoldCode, { name: HoldState; color: HoldColor; displayColor?: HoldColor }>
-> = {
-  kilter: {
-    // Product 1 – Kilter Board Original
-    12: { name: 'STARTING', color: '#00FF00' },
-    13: { name: 'HAND', color: '#00FFFF' },
-    14: { name: 'FINISH', color: '#FF00FF' },
-    15: { name: 'FOOT', color: '#FFAA00' },
-    // Product 2 – JUUL
-    20: { name: 'STARTING', color: '#00FF00' },
-    21: { name: 'HAND', color: '#00FFFF' },
-    22: { name: 'FINISH', color: '#FF00FF' },
-    23: { name: 'FOOT', color: '#FFA500' },
-    // Product 3 – Demo Board
-    24: { name: 'STARTING', color: '#00FF00' },
-    25: { name: 'HAND', color: '#00FFFF' },
-    26: { name: 'FINISH', color: '#FF00FF' },
-    27: { name: 'FOOT', color: '#FFA500' },
-    // Product 4 – BKB Board
-    28: { name: 'STARTING', color: '#00FF00' },
-    29: { name: 'HAND', color: '#00FFFF' },
-    30: { name: 'FINISH', color: '#FF00FF' },
-    31: { name: 'FOOT', color: '#FFA500' },
-    // Product 5 – Spire
-    32: { name: 'STARTING', color: '#00FF00' },
-    33: { name: 'HAND', color: '#00FFFF' },
-    34: { name: 'FINISH', color: '#FF00FF' },
-    35: { name: 'FOOT', color: '#FFA500' },
-    // Product 6 – Tycho (color mode, no start/finish semantics)
-    36: { name: 'HAND', color: '#00FFFF' },
-    37: { name: 'HAND', color: '#FF00FF' },
-    38: { name: 'HAND', color: '#FFFF00' },
-    39: { name: 'HAND', color: '#00FF00' },
-    40: { name: 'HAND', color: '#FF0000' },
-    41: { name: 'HAND', color: '#0000FF' },
-    // Product 7 – Kilter Board Homewall
-    42: { name: 'STARTING', color: '#00FF00' },
-    43: { name: 'HAND', color: '#00FFFF' },
-    44: { name: 'FINISH', color: '#FF00FF' },
-    45: { name: 'FOOT', color: '#FFAA00' },
-  },
-  tension: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-    5: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    6: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    7: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    8: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-  moonboard: {
-    42: { name: 'STARTING', color: '#00FF00', displayColor: '#44FF44' },
-    43: { name: 'HAND', color: '#0000FF', displayColor: '#4444FF' },
-    44: { name: 'FINISH', color: '#FF0000', displayColor: '#FF3333' },
-  },
-  decoy: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#0000FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-  touchstone: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-  grasshopper: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#4455FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-};
-
-// Warned hold states to avoid log spam
-const warnedHoldStates = new Set<string>();
-
-/**
- * Convert lit up holds string to a map of frames.
- * Each frame maps hold IDs to their state, color, and display color.
- */
-export function convertLitUpHoldsStringToMap(litUpHolds: string, board: BoardName): Record<number, LitUpHoldsMap> {
-  return litUpHolds
-    .split(',')
-    .filter((frame) => frame)
-    .reduce(
-      (frameMap, frameString, frameIndex) => {
-        const frameHoldsMap = Object.fromEntries(
-          frameString
-            .split('p')
-            .filter((hold) => hold)
-            .map((holdData) => holdData.split('r').map((str) => Number(str)))
-            .map(([holdId, stateCode]) => {
-              const stateInfo = HOLD_STATE_MAP[board]?.[stateCode];
-              if (!stateInfo) {
-                const warnKey = `${board}:${stateCode}`;
-                if (!warnedHoldStates.has(warnKey)) {
-                  warnedHoldStates.add(warnKey);
-                  console.warn(
-                    `HOLD_STATE_MAP is missing values for ${board} status code: ${stateCode} (this warning is only shown once per status code)`,
-                  );
-                }
-                return [holdId || 0, { state: `${holdId}=${stateCode}` as HoldState, color: '#FFF', displayColor: '#FFF' }];
-              }
-              const { name, color, displayColor } = stateInfo;
-              return [holdId, { state: name, color, displayColor: displayColor || color }];
-            }),
-        );
-        frameMap[frameIndex] = frameHoldsMap as LitUpHoldsMap;
-        return frameMap;
-      },
-      {} as Record<number, LitUpHoldsMap>,
-    );
-}
+// Re-export from the canonical source in @boardsesh/board-constants
+export { HOLD_STATE_MAP, convertLitUpHoldsStringToMap } from '@boardsesh/board-constants/hold-states';
