+++ packages/board-constants/src/index.ts
@@ -3,3 +3,5 @@ export type { BoardName } from '@boardsesh/shared-schema';
 export * from './types';
 export * from './product-sizes';
 export * from './led-placements';
+export * from './hold-states';
+export * from './grade-colors';
