+++ packages/web/app/components/board-renderer/types.ts
@@ -1,14 +1,15 @@
-import { BoardName } from '@/app/lib/types';
 import { MOONBOARD_ENABLED } from '@/app/lib/moonboard-config';
+import type { BoardName } from '@/app/lib/types';
+
+// Re-export hold state types and constants from the canonical source
+export { HOLD_STATE_MAP, convertLitUpHoldsStringToMap, STATE_TO_PRIMARY_CODE } from '@boardsesh/board-constants/hold-states';
+export type { HoldCode, HoldColor, HoldRenderStyle, HoldStateInfo } from '@boardsesh/board-constants/hold-states';
+export type { HoldState, LitupHold, LitUpHoldsMap } from '@boardsesh/shared-schema';
 
 export type LitUpHolds = string;
 
-export type HoldState = 'OFF' | 'STARTING' | 'FINISH' | 'HAND' | 'FOOT' | 'ANY' | 'NOT' | 'AUX';
 export type HoldsArray = Array<HoldRenderData>;
 
-export type HoldColor = string;
-export type HoldCode = number;
-export type HoldRenderStyle = 'circle' | 'above-marker';
 export type HoldRenderData = {
   id: number;
   mirroredHoldId: number | null;
@@ -16,8 +17,6 @@ export type HoldRenderData = {
   cy: number;
   r: number;
 };
-export type LitupHold = { state: HoldState; color: string; displayColor: string };
-export type LitUpHoldsMap = Record<HoldCode, LitupHold>;
 
 export interface HeatmapData {
   holdId: number;
@@ -28,8 +27,8 @@ export interface HeatmapData {
   footUses: number;
   finishUses: number;
   averageDifficulty: number | null;
-  userAscents?: number; // Added for user-specific ascent data
-  userAttempts?: number; // Added for user-specific attempt data
+  userAscents?: number;
+  userAttempts?: number;
 }
 
 /** Thumbnail render width in pixels. Covers 3x retina at ~64px CSS display. */
@@ -39,90 +38,3 @@ export const THUMBNAIL_WIDTH = 200;
 export const supported_boards: BoardName[] = MOONBOARD_ENABLED
   ? ['kilter', 'tension', 'moonboard', 'decoy', 'touchstone', 'grasshopper']
   : ['kilter', 'tension', 'decoy', 'touchstone', 'grasshopper'];
-
-// Mapping object for board-specific hold states
-export const HOLD_STATE_MAP: Record<
-  BoardName,
-  Record<HoldCode, { name: HoldState; color: HoldColor; displayColor?: HoldColor; renderStyle?: HoldRenderStyle }>
-> = {
-  kilter: {
-    // Product 1 – Kilter Board Original
-    12: { name: 'STARTING', color: '#00FF00' },
-    13: { name: 'HAND', color: '#00FFFF' },
-    14: { name: 'FINISH', color: '#FF00FF' },
-    15: { name: 'FOOT', color: '#FFAA00' },
-    // Product 2 – JUUL
-    20: { name: 'STARTING', color: '#00FF00' },
-    21: { name: 'HAND', color: '#00FFFF' },
-    22: { name: 'FINISH', color: '#FF00FF' },
-    23: { name: 'FOOT', color: '#FFA500' },
-    // Product 3 – Demo Board
-    24: { name: 'STARTING', color: '#00FF00' },
-    25: { name: 'HAND', color: '#00FFFF' },
-    26: { name: 'FINISH', color: '#FF00FF' },
-    27: { name: 'FOOT', color: '#FFA500' },
-    // Product 4 – BKB Board
-    28: { name: 'STARTING', color: '#00FF00' },
-    29: { name: 'HAND', color: '#00FFFF' },
-    30: { name: 'FINISH', color: '#FF00FF' },
-    31: { name: 'FOOT', color: '#FFA500' },
-    // Product 5 – Spire
-    32: { name: 'STARTING', color: '#00FF00' },
-    33: { name: 'HAND', color: '#00FFFF' },
-    34: { name: 'FINISH', color: '#FF00FF' },
-    35: { name: 'FOOT', color: '#FFA500' },
-    // Product 6 – Tycho (color mode, no start/finish semantics)
-    36: { name: 'HAND', color: '#00FFFF' },
-    37: { name: 'HAND', color: '#FF00FF' },
-    38: { name: 'HAND', color: '#FFFF00' },
-    39: { name: 'HAND', color: '#00FF00' },
-    40: { name: 'HAND', color: '#FF0000' },
-    41: { name: 'HAND', color: '#0000FF' },
-    // Product 7 – Kilter Board Homewall
-    42: { name: 'STARTING', color: '#00FF00' },
-    43: { name: 'HAND', color: '#00FFFF' },
-    44: { name: 'FINISH', color: '#FF00FF' },
-    45: { name: 'FOOT', color: '#FFAA00' },
-  },
-  tension: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-    5: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    6: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    7: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    8: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-  // MoonBoard hold states (no foot holds)
-  // Values 42-44 are used by saved MoonBoard climbs.
-  // Values 45-48 are additional live-BLE preview roles emitted by the ESP32 dev firmware.
-  moonboard: {
-    42: { name: 'STARTING', color: '#00FF00', displayColor: '#44FF44' }, // start
-    43: { name: 'HAND', color: '#0000FF', displayColor: '#4444FF' }, // right/plain hand
-    44: { name: 'FINISH', color: '#FF0000', displayColor: '#FF3333' }, // finish
-    45: { name: 'FOOT', color: '#00FFFF', displayColor: '#66F0FF' }, // foot
-    46: { name: 'AUX', color: '#FFE066', displayColor: '#FFE066', renderStyle: 'above-marker' }, // above-hold marker
-    47: { name: 'HAND', color: '#8B5CF6', displayColor: '#C084FC' }, // left hand
-    48: { name: 'HAND', color: '#FF4FA3', displayColor: '#FF7DBB' }, // match hand
-  },
-  // New Aurora boards use the same 1/2/3/4 role codes as Tension-style layouts.
-  decoy: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#0000FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-  touchstone: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#4444FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-  grasshopper: {
-    1: { name: 'STARTING', displayColor: '#00DD00', color: '#00FF00' },
-    2: { name: 'HAND', displayColor: '#4455FF', color: '#0000FF' },
-    3: { name: 'FINISH', displayColor: '#FF0000', color: '#FF0000' },
-    4: { name: 'FOOT', displayColor: '#FF00FF', color: '#FF00FF' },
-  },
-};
